moved files to "correct" locations
