+++ CruiseLink/com.csi3471.unbearables.maven.cruiselink/src/Controller/ReservationDatabaseController.java
@@ -1,5 +1,7 @@
-package person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.main.java.com.csi3471.unbearables.maven.cruiselink;
+package person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.Controller;
 
+import person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.Controller.cruiselink.Reservation;
+import person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.Repository.ReservationDatabase;
 import person.Guest;
 
 import java.util.ArrayList;
