+++ CruiseLink/com.csi3471.unbearables.maven.cruiselink/src/Controller/cruiselink/QualityLevel.java
@@ -1,4 +1,4 @@
-package person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.main.java.com.csi3471.unbearables.maven.cruiselink;
+package person.CruiseLink.com.csi3471.unbearables.maven.cruiselink.src.Controller.cruiselink;
 
 /* Team: UnBEARables
  * Course: CSI 3471
